Randomizing within categories now works
Stopped using stack overflow and created my own randomizing function
which turned out to be a lot easier than I was trying to make it. Next
step is to convert the choices in the quiz from radio boxes to check
boxes so I don’t have to deal with repopulation of previously selected
choices to questions.